% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{chpwexp}
\alias{chpwexp}
\title{Piecewise Exponential cumulative hazard function}
\usage{
chpwexp(x, rate = 1, t = 0)
}
\arguments{
\item{x, q}{:      vector of quantiles.}

\item{rate}{:    vector of rates.}

\item{t}{: vector of the same length as rate, giving the times at which the rate changes. The first element of t should be 0, and t should be in increasing order.}

\item{p}{:   vector of probabilities.}

\item{n}{:   number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{log}{:    logical; if TRUE, log function is returned.}
}
\value{
Piecewise Exponential cumulative hazard function
}
\description{
Piecewise Exponential cumulative hazard function
}
