% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{rsim.pwexp}
\alias{rsim.pwexp}
\title{Simulation from a Piecewise Exponential (piece-wise constant rate)}
\usage{
rsim.pwexp(seed, n, rate = 1, t = 0)
}
\arguments{
\item{seed}{: seed for simulation}

\item{n}{:   number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{rate}{:    vector of rates.}

\item{t}{: vector of the same length as rate, giving the times at which the rate changes. The first element of t should be 0, and t should be in increasing order.}

\item{x, q}{:      vector of quantiles.}

\item{p}{:   vector of probabilities.}

\item{log}{:    logical; if TRUE, log function is returned.}
}
\value{
generates random numbers from a Piecewise Exponential hazard function
}
\description{
Simulation from a Piecewise Exponential (piece-wise constant rate)
}
