% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{sim_pophaz}
\alias{sim_pophaz}
\title{Function to simulate survival times from the population hazard}
\usage{
sim_pophaz(seed, lst)
}
\arguments{
\item{lst}{: list with time points at which the piecewise hazard jumps (times) and the values of the piecewise hazard (hrates)}

\item{seed:}{seed for simulation}
}
\value{
a list containing the simulated survival times and the ID
}
\description{
Function to simulate survival times from the population hazard
}
