% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{simDesMatrix}
\alias{simDesMatrix}
\title{Function to simulate a design matrix. Based on real data on colon and lung (male and female) patients in England.}
\usage{
simDesMatrix(
  seed = 123,
  n = 100,
  admin.cens = "31-12-2015",
  scale.age = FALSE,
  site = "colon",
  sex = "female"
)
}
\arguments{
\item{seed}{: seed for simulation ("DD-MM-YYYY")}

\item{n}{: Sample size}

\item{admin.cens}{(year of administrative censoring),}

\item{site}{: cancer site ("lung" or "colon")}

\item{sex}{:  sex ("female", "male")}

\item{years.diag}{: years of diagnosis in the simulation (scalar or vector)}

\item{Returns:}{ID, sex (0 = male, 1 = female), dod (date of diagnosis), IMD (Index of Multiple Deprivation),}
}
\value{
a design matrix containing the ID, date.diag, dep, gor, age, and fup.time.
}
\description{
Function to simulate a design matrix. Based on real data on colon and lung (male and female) patients in England.
}
